Stock Prediction model using AMZN
this is a LSTM using this work https://www.youtube.com/watch?v=q_HS4s1L8UI using amazon as a benchmark. it's indexes by date and will otherwise try and fit for the model. should be a simple matter to plug in the KO data and get this whole thing working